update: used mousedown to close modal, added some error handling for user signin
