+++ store/reducer/navbar.js
@@ -1,0 +1,22 @@
+// import {} from '';
+
+// // initial State
+// const initialState = {};
+
+// const navbarReducer = (state = initialState, action) => {
+//   switch (action.type) {
+//     case SHOPPER: {
+//       return action.navbar;
+//     }
+//     case SELLER: {
+//       return action.navbar;
+//     }
+//     case COURIER: {
+//       return action.navbar;
+//     }
+//     default:
+//       return state;
+//   }
+// };
+
+// export default navbarReducer;
